RavenDB-24609 Address review feedback: GetPageSize, MessagesList partial, HasAttachments flag
- Use RequestHandler.GetPageSize(25) instead of manual GetIntValueQueryString
- Extract MessagesList struct to separate file via partial class
- Add HasAttachments flag to AiConversationMessagesResult
